Swagger support + refactoring

+++ review-service/src/main/java/com/reviews/service/impl/ReviewServiceImpl.java
@@ -1,5 +1,7 @@
 package com.reviews.service.impl;
 
+import static org.springframework.http.HttpStatus.NOT_FOUND;
+
 import com.reviews.entity.Review;
 import com.reviews.repository.ReviewRepository;
 import com.reviews.service.ReviewService;
@@ -8,6 +10,7 @@ import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
@@ -23,16 +26,32 @@ public class ReviewServiceImpl implements ReviewService {
 
   @Override
   public void updateReview(UUID reviewId, String content, int rating) {
+    boolean answer = reviewRepository.existsById(reviewId);
+    if(answer){
+      reviewRepository.deleteById(reviewId);
+    } else {
+      throw new ResponseStatusException(NOT_FOUND, "Such review does not exist");
+    }
     reviewRepository.setReviewInfoById(reviewId,content,rating);
   }
 
   @Override
   public void deleteReview(UUID reviewId) {
-    reviewRepository.deleteById(reviewId);
+    boolean answer = reviewRepository.existsById(reviewId);
+    if(answer){
+      reviewRepository.deleteById(reviewId);
+    } else {
+      throw new ResponseStatusException(NOT_FOUND, "Such review does not exist");
+    }
   }
 
   @Override
   public List<Review> getReviewsByProductId(UUID productId, int count) {
-    return reviewRepository.getReviewsLimited(productId,count);
+    boolean answer = reviewRepository.existsByProductId(productId);
+    if(answer) {
+      return reviewRepository.getReviewsLimited(productId, count);
+    } else {
+      throw new ResponseStatusException(NOT_FOUND, "No reviews for such product");
+    }
   }
 }
